version 1.1

+++ install.sh
@@ -1,5 +1,9 @@
 sudo apt-get install git python3 python3-pip -y
-rm -rf /home/script/hacksec-cli
+
+if [ -d ~/hacksec-cli ] 
+then
+    rm -rf /home/script/hacksec-cli
+fi
 git clone https://github.com/ScRiPt1337/hacksec-cli ~/hacksec-cli
 pip install -r  ~/hacksec-cli/requirements.txt
 echo "python3 ~/hacksec-cli/hacksec_cli/app.py" > ~/.local/bin/hacksec
